+++ AsBuiltReport.Microsoft.AD/Src/Public/Invoke-AsBuiltReport.Microsoft.AD.ps1
@@ -5,7 +5,7 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     .DESCRIPTION
         Documents the configuration of Microsoft AD in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        1.0.0
+        Version:        1.0.1
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -22,8 +22,14 @@ function Invoke-AsBuiltReport.Microsoft.AD {
         [PSCredential] $Credential
     )
 
-    #Requires -RunAsAdministrator
     #Requires -Version 7.4
+
+    $IsAdmin = ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+
+    if (-not $IsAdmin) {
+        Write-Error -Message $reportTranslate.InvokeAsBuiltReportMicrosoftAD.RunAsAdministrator
+        break
+    }
 
     if ($psISE) {
         Write-Error -Message $reportTranslate.InvokeAsBuiltReportMicrosoftAD.PwshISE
@@ -141,7 +147,11 @@ function Invoke-AsBuiltReport.Microsoft.AD {
             throw ($reportTranslate.InvokeAsBuiltReportMicrosoftAD.ForestError -f $System, $_.Exception.Message)
         }
 
-        $script:ForestInfo = $ADSystem.RootDomain.toUpper()
+        if (-not $ADSystem) {
+            throw ($reportTranslate.InvokeAsBuiltReportMicrosoftAD.ForestError -f $System, $reportTranslate.InvokeAsBuiltReportMicrosoftAD.NoData)
+        } else {
+            $script:ForestInfo = $ADSystem.RootDomain.toUpper()
+        }
         $RootDomains = $ADSystem.RootDomain
         $ChildDomains = [System.Collections.Generic.List[object]]::new()
         if ($Options.Include.Domains) {