+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##### This project is community maintained and has no sponsorship from Microsoft, its employees or any of its affiliates.
 
+## [1.0.1] - 2026-06-22
+
+### Added
+
+- Add cmdlet `Get-AbrAdLog` to collect report diagnostic log
+
+### Changed
+
+- Bump module version to `1.0.1`
+- Upgrade AsBuiltReport.Diagram module to version `1.0.8`
+- Upgrade AsBuiltReport.Chart module to version `0.3.3`
+- Set an option to enable/disable the main logo of the diagrams
+- Update github actions to latest releases
+
+### Fixed
+
+- Add validation for ADSystem before accessing RootDomain to prevent errors
+- Add success message logging for completed commands in `Invoke-CommandWithTimeout` function
+- Update localization files to include NoData messages and improve key matching in tests
+- Fix [255](https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD/issues/255)
+- Fix: remove non-printable ASCII characters from a string
+
 ## [1.0.0] - 2026-05-02
 
 ### Added